requirements.txt
fasttext==0.9.2
fr-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_lg-3.0.0/fr_core_news_lg-3.0.0-py3-none-any.whl
omegaconf==2.1.0
poutyne==1.6
spacy==3.0.6
torch==1.9.0
transformers==4.11.3
hydra-core
tensorboard
pandas
